+++ spring-05-books/src/main/java/ru/otus/spring05books/dao/AuthorDaoJdbc.java
@@ -1,44 +1,54 @@
 package ru.otus.spring05books.dao;
 
-import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import ru.otus.spring05books.domain.Author;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Класс AuthorDaoJdbc реализует интерфейс AuthorDao для JDBC
  */
 @Repository
 public class AuthorDaoJdbc implements AuthorDao {
-    private final JdbcOperations jdbc;
+    private final NamedParameterJdbcOperations jdbc;
 
     /**
      * Конструктор класса
      *
      * @param jdbc
      */
-    public AuthorDaoJdbc(JdbcOperations jdbc) {
+    public AuthorDaoJdbc(NamedParameterJdbcOperations jdbc) {
         this.jdbc = jdbc;
     }
 
     /**
      * Метод createAuthor создает нового автора в библиотеке
+     * Если такой автор (fullName) уже есть, то возвращается его id
      *
      * @param author
      * @return
+     * @see AuthorDaoJdbc#getIdByAuthor
      */
     @Override
     public long createAuthor(Author author) {
-        long id = getIdByAuthor(author);
-        if (id == 0) {
-            jdbc.update("insert into author (fullname) values (?)", author.getFullName());
-            id = getIdByAuthor(author);
+        long authorId = getIdByAuthor(author);
+        if (authorId == 0) {
+            MapSqlParameterSource params = new MapSqlParameterSource();
+            params.addValue("fullname", author.getFullName());
+            KeyHolder keyHolder = new GeneratedKeyHolder();
+            jdbc.update("insert into author (fullname) values (:fullname)", params, keyHolder);
+            return keyHolder.getKey().longValue();
+        } else {
+            return authorId;
         }
-        return id;
     }
 
     /**
@@ -49,7 +59,9 @@ public class AuthorDaoJdbc implements AuthorDao {
      */
     @Override
     public boolean updateAuthor(Author author) {
-        return false;
+        int result = jdbc.update("update author set fullname = :fullname where id = :id",
+                Map.of("id", author.getId(), "fullname", author.getFullName()));
+        return result == 1 ? true : false;
     }
 
     /**
@@ -60,7 +72,8 @@ public class AuthorDaoJdbc implements AuthorDao {
      */
     @Override
     public boolean deleteAuthor(Author author) {
-        return false;
+        int result = jdbc.update("delete from author where id = :id", Map.of("id", author.getId()));
+        return result == 1 ? true : false;
     }
 
     /**
@@ -71,7 +84,12 @@ public class AuthorDaoJdbc implements AuthorDao {
      */
     @Override
     public Author getAuthorById(long id) {
-        return null;
+        List<Author> authorList = jdbc.query("select id, fullname " +
+                        "from author " +
+                        "where id = :id",
+                Map.of("id", id),
+                new AuthorMapper());
+        return authorList.size() == 0 ? null : authorList.get(0);
     }
 
     /**
@@ -84,9 +102,9 @@ public class AuthorDaoJdbc implements AuthorDao {
      */
     @Override
     public long getIdByAuthor(Author author) {
-        List<Long> listId;
-        listId = jdbc.query("select id from author where fullname = ?", new IdMapper(), author.getFullName());
-        return listId.size() == 0 ? 0 : listId.get(0);
+        List<Author> authorList = jdbc.query("select id, fullname " + "from author " + "where fullname = :fullname",
+                Map.of("fullname", author.getFullName()), new AuthorMapper());
+        return authorList.size() == 0 ? 0 : authorList.get(0).getId();
     }
 
     /**
@@ -96,7 +114,8 @@ public class AuthorDaoJdbc implements AuthorDao {
      */
     @Override
     public List<Author> getAllAuthors() {
-        return null;
+        List<Author> authorList = jdbc.query("select id, fullname from author", Map.of("", ""), new AuthorMapper());
+        return authorList;
     }
 
     /**
@@ -106,17 +125,17 @@ public class AuthorDaoJdbc implements AuthorDao {
      */
     @Override
     public int getCountOfAuthors() {
-        Integer count = jdbc.queryForObject("select count(*) from author", Integer.class);
-        return count == null? 0: count;
+        Integer count = jdbc.getJdbcOperations().queryForObject("select count(*) from author", Integer.class);
+        return count == null ? 0 : count;
     }
 
     /**
-     * Класс IdMapper формирует набор для получаемого результата из jdbc.query
+     * AuthorMapper - результирующий запрос для Автора: id, fullname
      */
-    private static class IdMapper implements RowMapper<Long> {
+    private static class AuthorMapper implements RowMapper<Author> {
         @Override
-        public Long mapRow(ResultSet rs, int rowNum) throws SQLException {
-            return new Long(rs.getLong("id"));
+        public Author mapRow(ResultSet rs, int rowNum) throws SQLException {
+            return new Author(rs.getLong("id"), rs.getString("fullname"));
         }
     }
