+++ spring-05-books/src/main/java/ru/otus/spring05books/dao/GenreDaoJdbc.java
@@ -1,27 +1,31 @@
 package ru.otus.spring05books.dao;
 
-import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import ru.otus.spring05books.domain.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Класс GenreDaoJdbc реализует интерфейс GenreDao для JDBC
  */
 @Repository
 public class GenreDaoJdbc implements GenreDao {
-    private final JdbcOperations jdbc;
+    private final NamedParameterJdbcOperations jdbc;
 
     /**
      * Конструктор класса
      *
      * @param jdbc
      */
-    public GenreDaoJdbc(JdbcOperations jdbc) {
+    public GenreDaoJdbc(NamedParameterJdbcOperations jdbc) {
         this.jdbc = jdbc;
     }
 
@@ -33,12 +37,16 @@ public class GenreDaoJdbc implements GenreDao {
      */
     @Override
     public long createGenre(Genre genre) {
-        Long id = getIdByGenre(genre);
-        if (id == 0) {
-            jdbc.update("insert into genre(name) values(?)", genre.getName());
-            id = getIdByGenre(genre);
+        long genreId = getIdByGenre(genre);
+        if (genreId == 0) {
+            MapSqlParameterSource params = new MapSqlParameterSource();
+            params.addValue("name", genre.getName());
+            KeyHolder keyHolder = new GeneratedKeyHolder();
+            jdbc.update("insert into genre (name) values (:name)", params, keyHolder);
+            return keyHolder.getKey().longValue();
+        } else {
+            return genreId;
         }
-        return id;
     }
 
     /**
@@ -49,7 +57,9 @@ public class GenreDaoJdbc implements GenreDao {
      */
     @Override
     public boolean updateGenre(Genre genre) {
-        return false;
+        int result = jdbc.update("update genre set name = :name where id = :id",
+                Map.of("id", genre.getId(), "name", genre.getName()));
+        return result == 1 ? true : false;
     }
 
     /**
@@ -60,7 +70,8 @@ public class GenreDaoJdbc implements GenreDao {
      */
     @Override
     public boolean deleteGenre(Genre genre) {
-        return false;
+        int result = jdbc.update("delete from genre where id = :id", Map.of("id", genre.getId()));
+        return result == 1 ? true : false;
     }
 
     /**
@@ -71,22 +82,24 @@ public class GenreDaoJdbc implements GenreDao {
      */
     @Override
     public Genre getGenreById(long id) {
-
-        return null;
+        List<Genre> genreList = jdbc.query("select id, name from genre where id = :id", Map.of("id", id),
+                new GenreMapper());
+        return genreList.size() == 0 ? null : genreList.get(0);
     }
 
     /**
      * Метод getIdByGenre возвращает id передаваемого жанра
      * Метод query использует jdbc.query, так как возможно получение как 0, так и 1, а jdbc.queryForObject выбрасывает
      * исключение если результат 0
+     *
      * @param genre
      * @return
      */
     @Override
     public long getIdByGenre(Genre genre) {
-        List<Long> listId;
-        listId = jdbc.query("select id from genre where name = ?", new IdMapper(), genre.getName());
-        return listId.size() == 0 ? 0 : listId.get(0);
+        List<Genre> genreList = jdbc.query("select id, name " + "from genre " + "where name = :name",
+                Map.of("name", genre.getName()), new GenreMapper());
+        return genreList.size() == 0 ? 0 : genreList.get(0).getId();
     }
 
     /**
@@ -96,7 +109,8 @@ public class GenreDaoJdbc implements GenreDao {
      */
     @Override
     public List<Genre> getAllGenres() {
-        return null;
+        List<Genre> genreList = jdbc.query("select id, name from genre", Map.of("", ""), new GenreMapper());
+        return genreList;
     }
 
     /**
@@ -106,17 +120,18 @@ public class GenreDaoJdbc implements GenreDao {
      */
     @Override
     public int getCountOfGenres() {
-        Integer count = jdbc.queryForObject("select count(*) from genre", Integer.class);
-        return count == null? 0: count;
+        Integer count = jdbc.getJdbcOperations().queryForObject("select count(*) from genre", Integer.class);
+        return count == null ? 0 : count;
     }
 
     /**
-     * Класс IdMapper формирует набор для получаемого результата из jdbc.query
+     * GenreMapper - результирующий запрос для Жанра: id, fullname
      */
-    private static class IdMapper implements RowMapper<Long> {
+    private static class GenreMapper implements RowMapper<Genre> {
         @Override
-        public Long mapRow(ResultSet rs, int rowNum) throws SQLException {
-            return new Long(rs.getLong("id"));
+        public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
+            return new Genre(rs.getLong("id"), rs.getString("name"));
         }
     }
+
 }
